frontend: Move idian widgets to shared/qt
