+++ frontend/cmake/ui-dialogs.cmake
@@ -1,3 +1,9 @@
+if(NOT TARGET OBS::idian)
+  add_subdirectory("${CMAKE_SOURCE_DIR}/shared/qt/idian" "${CMAKE_BINARY_DIR}/shared/qt/idian")
+endif()
+
+target_link_libraries(obs-studio PRIVATE OBS::idian)
+
 if(NOT TARGET OBS::properties-view)
   add_subdirectory("${CMAKE_SOURCE_DIR}/shared/properties-view" "${CMAKE_BINARY_DIR}/shared/properties-view")
 endif()
