+++ shared/qt/idian/OBSCheckBox.cpp
