+++ shared/qt/idian/OBSDoubleSpinBox.cpp
