+++ shared/qt/idian/OBSDoubleSpinBox.hpp
