+++ shared/qt/idian/OBSGroupBox.cpp
