+++ shared/qt/idian/OBSIdianWidget.hpp
