+++ shared/qt/idian/OBSPropertiesList.cpp
